celery_test/requirements.txt
aiohttp==3.9.1
aioredis==2.0.1
celery==5.3.6
fastapi==0.104.1
prisma==0.11.0
pydantic==2.5.2
python-dotenv==1.0.0
redis==5.0.1
uvicorn==0.24.0.post1